+++ src/main/java/com/pickpick/server/service/PhotoService.java
@@ -1,6 +1,7 @@
 package com.pickpick.server.service;
 
 import com.pickpick.server.apiPayload.code.status.ErrorStatus;
+import com.pickpick.server.apiPayload.exception.handler.PhotoHandler;
 import com.pickpick.server.apiPayload.exception.handler.UserHandler;
 import com.pickpick.server.domain.Category;
 import com.pickpick.server.domain.Photo;
@@ -12,13 +13,14 @@ import com.pickpick.server.repository.UsersRepository;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
+import javax.swing.text.html.Option;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 @Service
 @RequiredArgsConstructor
-@Transactional(readOnly = true)
+@Transactional
 public class PhotoService {
 
     private final PhotoRepository photoRepository;
@@ -26,27 +28,34 @@ public class PhotoService {
     private final UsersRepository usersRepository;
 
     public Photo create(PhotoRequest.CreateDTO request) {
+        Optional<Users> user = usersRepository.findById(request.getUserId());
+        if(user.isEmpty()){
+            throw new UserHandler(ErrorStatus.MEMBER_NOT_FOUND);
+        }
         Photo photo = Photo.builder()
-            .user(usersRepository.findById(request.getUserId()).get())
+            .user(user.get())
             .imgUrl(request.getImgUrl())
             .build();
         return photoRepository.save(photo);
     }
 
     public Photo createCategory(PhotoRequest.CreateCategoryDTO request) {
-        Photo photo = photoRepository.findById(request.getPhotoId()).orElseThrow();
-        List<Category> categories = categoryRepository.findByPhoto(photo);
+        Optional<Photo> photo = photoRepository.findById(request.getPhotoId());
+        if(photo.isEmpty()){
+            throw new PhotoHandler(ErrorStatus.PHOTO_NOT_FOUND);
+        }
+        List<Category> categories = categoryRepository.findByPhoto(photo.get());
 
         categoryRepository.deleteAll(categories);
 
         for (String category : request.getCategoryList()) {
             Category newCategory = Category.builder()
-                .photo(photo)
+                .photo(photo.get())
                 .name(category)
                 .build();
             categoryRepository.save(newCategory);
         }
-        return photoRepository.save(photo);
+        return photoRepository.save(photo.get());
     }
 
     public List<String> getPhotos(String email) {
